requirements.txt
argon2-cffi==25.1.0
cryptography==46.0.3
mysql-connector-python==9.5.0
requests==2.32.5
rich==14.2.0
zxcvbn==4.5.0
pyperclip==1.11.0